nice refactor
